+++ R/vsearch.R
@@ -550,6 +550,7 @@ chimera_removal_vs <-
              'Select_only_non_chim', or 'Select_only_chim'"
         )
       }
+      seq_tab_final <- t(data.frame(seq_tab_final))
       return(seq_tab_final)
     } else if (inherits(object, "phyloseq")) {
       if (sum(taxa_sums(object) == 0) > 0) {